+++ packages/rfc-client-soap-xml/package.json
@@ -14,6 +14,7 @@
     "typescript": "^4.5.5"
   },
   "dependencies": {
+    "@abaplint/runtime": "^1.7.29",
     "cross-fetch": "^3.1.5",
     "fast-xml-parser": "^4.0.1"
   }
